refactor(mcp): decompose jsonrpc + feature-gate GenericTauriBackend (#1200)
- Split src/jsonrpc.rs (385 LOC) into jsonrpc/{mod,message,dispatch}.rs.
  message owns Request/Response/RpcError; dispatch owns MethodHandler/
  ResponseWriter/write_response/serve. mod.rs re-exports both flat so all
  jsonrpc::* public paths are unchanged.
- Move GenericTauriBackend (the always-Unimplemented placeholder) behind the
  off-by-default generic-tauri-backend cargo feature so the default public
  surface ships no always-Unimplemented backend. No unimplemented!() added;
  default-build behaviour unchanged (the type was wired into no binary).
- Update TODO.md (TD-002/TD-003 Open->Done), README.md, AGENTS.md.

Behaviour-preserving; architecture_fitness + clippy (feature on/off) green.
