+++ parish/crates/parish-mcp/src/jsonrpc/dispatch.rs
@@ -1,13 +1,9 @@
-//! Minimal line-delimited JSON-RPC 2.0 over async streams.
+//! Transport / dispatch layer for line-delimited JSON-RPC 2.0.
 //!
-//! MCP supports several transports; the most common (and the only one
-//! Claude Desktop / Claude Code currently configure) is **stdio with
-//! newline-delimited JSON messages**. This module owns the wire layer
-//! and is deliberately ignorant of MCP method names — that lives in
-//! [`crate::mcp`].
-//!
-//! Errors are surfaced via JSON-RPC error objects rather than panics so a
-//! malformed request from the client never tears down the server loop.
+//! Owns the [`MethodHandler`] trait, the [`ResponseWriter`], the
+//! `write_response` helper, and the [`serve`] read loop. Split out of the
+//! monolithic `jsonrpc` module (#1200) so the framing/dispatch logic is
+//! separable from the protocol value types in [`super::message`].
 //!
 //! Concurrency model: requests are processed serially in the order they
 //! arrive. JSON-RPC clients correlate by `id` so out-of-order responses
@@ -18,109 +14,12 @@
 //! semantics predictable and avoids losing responses when stdin closes
 //! while handlers are still in flight.
 
-use serde::{Deserialize, Deserializer, Serialize};
 use serde_json::Value;
 use std::sync::Arc;
 use tokio::io::{AsyncBufReadExt, AsyncWrite, AsyncWriteExt, BufReader};
 use tokio::sync::Mutex;
 
-/// JSON-RPC 2.0 request / notification.
-///
-/// `id` is `None` for notifications. We accept any JSON value for `id`
-/// (string, number, null) per the spec and echo it back verbatim.
-#[derive(Debug, Clone, Deserialize)]
-pub struct Request {
-    #[allow(dead_code)] // verified during deserialisation; kept for completeness
-    pub jsonrpc: String,
-    pub method: String,
-    #[serde(default)]
-    pub params: Value,
-    #[serde(default, deserialize_with = "deserialize_optional_id")]
-    pub id: Option<Value>,
-}
-
-/// JSON-RPC 2.0 response. Either `result` or `error` is set, never both.
-#[derive(Debug, Clone, Serialize)]
-pub struct Response {
-    pub jsonrpc: &'static str,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<RpcError>,
-    pub id: Value,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RpcError {
-    pub code: i64,
-    pub message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Value>,
-}
-
-fn deserialize_optional_id<'de, D>(deserializer: D) -> Result<Option<Value>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    Value::deserialize(deserializer).map(Some)
-}
-
-impl RpcError {
-    pub fn parse_error(detail: impl Into<String>) -> Self {
-        Self {
-            code: -32700,
-            message: detail.into(),
-            data: None,
-        }
-    }
-    pub fn invalid_request(detail: impl Into<String>) -> Self {
-        Self {
-            code: -32600,
-            message: detail.into(),
-            data: None,
-        }
-    }
-    pub fn method_not_found(method: &str) -> Self {
-        Self {
-            code: -32601,
-            message: format!("method not found: {method}"),
-            data: None,
-        }
-    }
-    pub fn invalid_params(detail: impl Into<String>) -> Self {
-        Self {
-            code: -32602,
-            message: detail.into(),
-            data: None,
-        }
-    }
-    pub fn internal(detail: impl Into<String>) -> Self {
-        Self {
-            code: -32603,
-            message: detail.into(),
-            data: None,
-        }
-    }
-}
-
-impl Response {
-    pub fn ok(id: Value, result: Value) -> Self {
-        Self {
-            jsonrpc: "2.0",
-            result: Some(result),
-            error: None,
-            id,
-        }
-    }
-    pub fn err(id: Value, error: RpcError) -> Self {
-        Self {
-            jsonrpc: "2.0",
-            result: None,
-            error: Some(error),
-            id,
-        }
-    }
-}
+use super::message::{Request, Response, RpcError};
 
 /// A handler is given the parsed `params` and must return either a
 /// JSON-encodable result or a typed [`RpcError`].
@@ -224,24 +123,6 @@ where
 mod tests {
     use super::*;
 
-    #[test]
-    fn response_serialises_only_result_when_ok() {
-        let resp = Response::ok(Value::from(7), serde_json::json!({"hello": "world"}));
-        let s = serde_json::to_string(&resp).unwrap();
-        assert!(s.contains("\"result\""));
-        assert!(!s.contains("\"error\""));
-        assert!(s.contains("\"id\":7"));
-    }
-
-    #[test]
-    fn response_serialises_only_error_when_err() {
-        let resp = Response::err(Value::from("abc"), RpcError::method_not_found("foo"));
-        let s = serde_json::to_string(&resp).unwrap();
-        assert!(s.contains("\"error\""));
-        assert!(!s.contains("\"result\""));
-        assert!(s.contains("method not found: foo"));
-    }
-
     /// Smoke test for the full read/dispatch/write loop using in-memory pipes.
     #[tokio::test]
     async fn serve_round_trips_a_request() {
@@ -258,7 +139,7 @@ mod tests {
 
         let handler = Arc::new(Echo);
         let writer = ResponseWriter::new(server_out);
-        let serve_task = tokio::spawn(super::serve(server_in, writer, handler));
+        let serve_task = tokio::spawn(serve(server_in, writer, handler));
 
         let req = b"{\"jsonrpc\":\"2.0\",\"method\":\"ping\",\"params\":{\"a\":1},\"id\":42}\n";
         client_in.write_all(req).await.unwrap();
@@ -293,7 +174,7 @@ mod tests {
         let (server_out, mut client_out) = tokio::io::duplex(4096);
 
         let writer = ResponseWriter::new(server_out);
-        let serve_task = tokio::spawn(super::serve(server_in, writer, Arc::new(OkHandler)));
+        let serve_task = tokio::spawn(serve(server_in, writer, Arc::new(OkHandler)));
 
         client_in
             .write_all(
@@ -332,7 +213,7 @@ mod tests {
         let (mut client_in, server_in) = tokio::io::duplex(4096);
         let (server_out, mut client_out) = tokio::io::duplex(4096);
         let writer = ResponseWriter::new(server_out);
-        let serve_task = tokio::spawn(super::serve(server_in, writer, Arc::new(Sink)));
+        let serve_task = tokio::spawn(serve(server_in, writer, Arc::new(Sink)));
 
         client_in
             .write_all(b"{\"jsonrpc\":\"2.0\",\"method\":\"shouldNotReply\"}\n")
@@ -366,7 +247,7 @@ mod tests {
         let (mut client_in, server_in) = tokio::io::duplex(4096);
         let (server_out, mut client_out) = tokio::io::duplex(4096);
         let writer = ResponseWriter::new(server_out);
-        let serve_task = tokio::spawn(super::serve(server_in, writer, Arc::new(Sink)));
+        let serve_task = tokio::spawn(serve(server_in, writer, Arc::new(Sink)));
 
         client_in.write_all(b"{not json}\n").await.unwrap();
         client_in.shutdown().await.unwrap();
